+++ src/Crate/PDO/PDOImplementationPhp7.php
@@ -20,35 +20,28 @@
  * software solely pursuant to the terms of the relevant commercial agreement.
  */
 
-/*
- * PDO compatibility for PHP 8.x.
- *
- * Reason:
- *  - https://github.com/php/php-src/pull/6220
- *
- * Implementation derived and inspired from:
- *  - https://github.com/doctrine/dbal/pull/4347
- *  - https://www.drupal.org/project/drupal/issues/3109885
- *  - https://www.drupal.org/project/drupal/issues/3156881
- *  - https://www.drupal.org/node/3170913
- *
- */
-
 declare(strict_types=1);
 
 namespace Crate\PDO;
 
-use const PHP_VERSION_ID;
+use function func_get_args;
 
 /**
- * Interface PDOInterface
- *
- * Used for unit testing the PDOStatement
- *
  * @internal
  */
-if (PHP_VERSION_ID >= 80000) {
-    class_alias('\Crate\PDO\PDOInterfacePhp8', '\Crate\PDO\PDOInterface');
-} else {
-    class_alias('\Crate\PDO\PDOInterfacePhp7', '\Crate\PDO\PDOInterface');
+trait PDOImplementationPhp7
+{
+    /**
+     * @return PDOStatement
+     */
+    public function query()
+    {
+        return $this->doQuery(...func_get_args());
+    }
+
+    /**
+     * @param $statement
+     * @return PDOStatement
+     */
+    abstract public function doQuery($statement);
 }